feat(TDI-45261) : move local sites validation in common
